Added the possibility to change the resolution on Android
A requestedSizeIndex of 0 returns on Android the default size of
640x480. If this size is not available the median element of the
available sizes array is returned.
An index of -n return the n-th smaller one.
An index of +n return the n-th bigger one.
This requestedSizeIndex has only an influence right now on Android.
On desktop systen OpenCV gets the max. available resolution.
On iOS the resolution is hardcoded to 640 x 480.*/