Hacked together the first version that can render images to the HMD
Added the SLVRCompositor for creating framebuffers for both eyes and outsourced conversions to SLVRConvert. There is now a new projection type for OpenVR which is creatively called "OpenVR". Nothing is rendered to the window when using this projection type to save performance. Speaking of which, the performance is currently very bad.