Add support for Clang on Windows
It's now possible to compile SLProject using clang-cl on Windows. clang-cl is supposed to be a "drop-in replacement" for the MSVC compiler, but there were still some flags that needed to be set. In addition, the CMakeLists were updated to account for a Clang compiler with an MSVC frontend (as opposed to a GNU frontend)