Save day, add pause particle, add correction gap, adapt with frustum culling., comments
Lot of thing:
Added button reset and pause/resume
When particle system pause the particle system is frozen, it doesn't update anymore, Adapt this with frustum culling.
Added correction gap which allow us to remove flickering of the recycling of particle with no velocity when disable, but this add a bug that is not there if enable of delay/lag/gap when the particles spawned (see documentation for more details).
Added comments in draw for frustum culling and paused