Demo scene with 3D box
Assets are now accessed using the SLAssetStore interface. SLAssetStore
uses the file system as a backend on most platforms, but when compiling
with Emscripten, it stores all files in memory. This is because handling
files in browsers is not that easy.

Many features of OpenCV (such as imread) are disabled when
compiling using Emscripten because they aren't available on this
platform.