[Emscripten] Fix transparent textures
stb_image always loaded images in RGBA format, which caused the renderer to disable depth testing for all textures. This led to rendering issues that we fixed temporarily by forcing depth testing to be always enabled.

Now images are loaded in their format on disk, so we can remove this hack.